+++ apps/gatherer/server/src/app.module.ts
@@ -1,7 +1,11 @@
 import { Module } from "@nestjs/common";
+import { CollectService } from "./collect.service";
 import { HealthController } from "./health.controller";
+import { SourceController } from "./source.controller";
+import { SourceService } from "./source.service";
 
 @Module({
-  controllers: [HealthController],
+  controllers: [HealthController, SourceController],
+  providers: [SourceService, CollectService],
 })
 export class AppModule {}