+++ infra/postgres-init.sql
@@ -2,3 +2,4 @@ CREATE DATABASE applicant;
 CREATE DATABASE investigator;
 CREATE DATABASE assessor;
 CREATE DATABASE manager;
+CREATE DATABASE gatherer;